theme: Create "small_header" for mobile and fix date/tags clash
On narrow screens the date and tags in post-meta were touching.
Ensure space between tags that are underneath each other by adding
a bottom margin to the tags anchors. This also ensures a bit of
spacing between the last row of tags and the date (which will
be floating under each other at this point).

I'm keeping floats, instead of flex, because these should not be
separate columns. The tags can take up full width if needed
and the date is just the below that. This makes better use of
vertical space.

Lastly, because floats require we clear them (with overflow:hidden),
this means the newly added margin-bottom won't "collapse" into
the margin of the post paragraphs. As such, manually undo the
the bottom margin of the last row of tags with a negative margin.