theme: Add md-center-box. Shrink bold. Collapse blockquote margin.
New md-center-box style that provides a light shadow and border so
that assets with a white background (as most of mine do) stand out
a bit more. Even more so for those with a slightly shaded background,
it conflicts less this way, by giving it a harder stop.

Make bolded text render at effectively the same size as regular text.
To so do, the letter spacing and font size have to be reduced.

Blockquote inner paragraphs were adding margin in addition to the
padding the blockquote had already, this was making it a bit too
spacious. This is prep for a post that will include a Twitter thread.