theme: Re-add <meta> description
I removed this in in favour of "og:description", but it seems
Google doesn't consider this as fallback for its magic snippet
generation, it only falls back to the legacy description field.
This seems wasteful but compression should make this relatively
cheap.

Discovered via the SEO audit of Google's PageSpeed Insights
and Chrome Lighthouse.

Also use 'site.description' as fallback on the home page.