theme: make images expand to full-width blocks on mobile
This was already done for <figure> but not for simple images.

For figure, the strategy remains the same. There is negative margins
to counter the page-level padding, and the content expands by itself
as space allows given max-width 100% will correspond to the width
of the widened <figure>.

For other wrappers this behaviour can be applied via class=md-block.

For simple images in an implicit paragraph (no figure, no md-block),
we also need to increase the max-width from 100% to calc(100% + margin),
because merely applying margin-left:-20px will just move it without
it taking up more space because the <p> wrapper is still the same
width as without widening.