theme: make tooltips accessible via focus tap on (i)
For <abbr>, <acronym> and other non-link elements this is based
purely on CSS and the HTML element with 'title'  and (focussable)
'tabindex' attributes. There is a 'before' pseudo-element that
generates the (i) button, which if focussed, displays the
'after' pseudo-element with the title value in a fixed overlay.

On desktop and tablets, the overlay is on the line below the (i)
taget, using position-absolute without any top/bottom set, which
it defaults to the relative static position which  we widen to
the full width and then move it down relatively using margin-top.

On mobile, the overlay is pinned to the bottom of the screen
instead.

For <a href>, we can't easily do this because clicking on the
'before' element would unavoidably cause the link itself also
to be clicked and lead to unintended navigation. I tried to
make this work by setting 'pointer-events: none' on the 'before'
element so that perhaps once focussed, the chain of
hover->focus->click would stop, but no dice. I think the order
might be hover->click->focus instead, which means by the time
focussed happened the click has already been registered.
The other problem is that when you set 'pointer-events: none' you
don't actually prevent pointer events, rather this makes the
target element invisible/transparent for pointer event purposes,
meaning that surrounding/underlying element receives the event
instead, which presents the same problem (link getting clicked).

As a solution, I require a child <span> to exist within the link,
which is automatically inserted server-side by a Markdown filter.
There is one line of JS code that uses a single global delegate
click handler to prevent propagation of clicks from <span title>
to <a title>.