theme: change `hr` style from pull out to regular content
Follows-up 20336ae, which treated `hr` as pull out full-width
elements. I'm reverting this because it makes it feels too much
like a top-level divider, as if the page has ended with an unrelated
article following after. It also didn't match the width of the
post metadata line under the heading (and couldn't) which made
looked off.