theme: remove need for <hr> between post meta and body
I tried removing this before but kept it because it appeared
to be what made Safari Reader ignore `.post-meta` (e.g. the bullet
point and post category links). Without it, Reader would correctly
hoist the `H1` as title,  and `time` as date, but then leave the
remainder behind as non-sensical first paragraph of the article,
e.g.

> Post title
> By Post author on Post date.
>
>
> • Some tag
> The actual first paragraph.

Notice the dangling bullet and random tag link.

Looking around the web for unofficial guides, I noticed many
of them used class=byline on their metadata line. I didn't use
this before because the semantic <h1> and <time> sufficed for it
to correctly hoist it from inside <article>, and it doesn't make
any sense to me that it would hoist those while leaving the first
of that now-incomplete line of text behind.

The fact that there was an <hr> after "Some tag" made Safari consider
the start of the real content after it.

Use class=byline to make this work instead.

I tried adding class=byline to <header class=post-meta> instead
of adding a new element, but Reader didn't like that either.