theme: switch from customized solarized-light to zenburn
I originally used the solarized-light theme (for Jekyll/Pygments
syntax highlighting) when the site was primarily light mode.

For dark mode I used a CSS filter() to make the colors brighter
(for use on a dark background, instead of light grey), and some
other filters to taste.

I never quite liked how it turned out, but it was a compromise to
not send two copies of the styles. Now that I've made dark mode
the default and only theme, I'd rather use a nicer theme designed
for use on a dark background.