theme: Switch syntax highlight from zenburn to monokai
I originally picked this when the site used light colors, which
I then dimmed down and inverted for dark mode. This wasn't a great
fit but avoided a more complex and larger payload for something I
didn't consider important enough to pay that for.

Now that the site is dark-only, switch to a color theme I actually
like and is tailored for use on a dark background.

Incidentally, when compressed, this happens to also be 0.5K smaller.